Replace seL4RPC with seL4RPCCall
In order to make the CAmkES repo work with the new-rt kernel,
we update all seL4RPC connectors to be seL4RPCCall.

The reason for this is that we haven't updated the seL4RPC connector
templates, since they are going to be deprecated soon, ostensibly.

Since we don't want all the CAmkES examples to just not work on the
new-rt kernel, we just update all the examples to use seL4RPCCall right
now.