Rename tcl includes to .inc
The python test scripts were treating every file in the tests folder
as a test if it ended in .tcl and had executable permissions.
The permission check was required because because there were tcl files that are only used as includes in other tests.
This meant that if you made a test and forgot to mark the file as
executable, the test would never run. Renaming the tcl include files to
.inc and removing the permission check prevents tests from being
accidentally excluded.