Legal C constant for static_assert
Despite the 'const' attribute a 'const' variable is still not considered a compile time
constant and so cannot be used in a static_assert. The code previously compiled as these
static_assert's were in `if (0)` blocks and whem compiled with optimisations the code
is elided before a compilation error can occur. When compiling at a lower optimization
level a compilation error is triggered.