Mark `inline` function as `static`
In c99 a function that is just `inline` needs an `extern` declaration somewhere so that
the compiler knows where to generate the non inlined version of the function for linking.
As this function has no `extern` declaration the compiler, depending on optimisation
levels, may emit a call to the function instead of inline, which will result in a link
time failure as none is generated.