Be more careful with argument lists
When I recently changed init-build.sh to actually quote "$@" and get
proper shell list semantics on the argument list, I exposed a bug in
our Tcl-based argument list construction in CAmkES.

tests/test.inc: Use [concat ...] instead of escaped spaces to construct
$cmake_props, the argument list that will be passed to init-build.sh.
This also keeps -DCAMKES_APP=$appname from having a space forciby
appended to it even if $extra_cmake_args is empty (which it often is);
that was the immediate cause of the build failures.

tests/build-cmake.inc: Use {*} to expand $cmake_props, which forces it
to be interpted as an argument _list_ instead of one giant string.

Thanks to Kent McLeod for working with me to uncover the mysteries of
Tcl.  And to Damian Conway for teaching me years ago that in Perl, the
fate of the universe hangs on the difference between lists and scalars.
It seems every scripting language struggles with this.