CMake: Import rumprun as sub project in apps
Rumprun's CMakeLists.txt declares target properties that are referenced
by camkes apps but aren't generator expressions. This means that the
rumprun CMakeLists.txt has to be processed before the camkes scripts.
Explicitly adding it as a subproject enables this and is safe in this
simplified app context where multiple sibling projects don't rely on
rumprun targets.