Updated serialserver_loopback to new libvirtqueue
The new libvirtqueue has scatter/gather capabilities as well as multiple
emtries in the ring. This update allows the loopback client to function
with the updated version of the serial server, and leverages the scatter
gather capability, however due to the design that was too specific to
the previous libvirtqueue, it wasn't possible to make use of the
multiple ring entries to make it truly asynchronous. More work would be
required to buffer the reads while the serial server is still writing,
which would make it substantially faster.