mcs-scheduling: Example to demo MCS sched logic
Consider 3 threads with budget and period setting as T_1(90 / 200),
T_2(60 / 250), and T_3(10 / 300), sorted in descending order of
priority. The sum of CPU utilization is 72.3% thus meet the
schedulability requirement of Rate-monotonic scheduling, which
means the lower priority thread T_3 will always have its chance to
run in any of its periods. The example demonstrates this working.